remove most transform pipelines
